+++ types/index.d.ts
@@ -1,0 +1,2 @@
+export * as events from "./events";
+export * as protyle from "./protyle";